penumbra_sdk_tct
Module error

Errors that can occur when inserting into a Tree, deserializing Proofs, or checking internal invariants.

Modules

block

Errors for block builders.

epoch

Errors for epoch builders.

proof

Errors from deserializing or verifying inclusion proofs.

Structs

InsertEpochError

The Tree was full when trying to insert an epoch into it.

RootDecodeError

An error occurred when decoding a tree root from bytes.

Enums

InsertBlockError

An error occurred when trying to insert a block into the Tree.

InsertError

An error occurred when trying to insert a [Commitment] into a Tree.