penumbra_sdk_tct::internal::complete
Trait ForgetOwned

pub trait ForgetOwned: Height + Sized {
    // Required method
    fn forget_owned(
        self,
        forgotten: Option<Forgotten>,
        index: impl Into<u64>,
    ) -> (Insert<Self>, bool);
}

Available on crate feature internal only.

Forget about the authentication path to a given index, when forgetting can turn the entirety of Self into a hash.

Required Methods

fn forget_owned( self, forgotten: Option<Forgotten>, index: impl Into<u64>, ) -> (Insert<Self>, bool)

Remove the witness for the given index and summarize the item as a single Hash if it now contains no more witnesses. If a forgotten version is specified, update the path down to the forgotten item to that version plus one.

Returns either (Self, boool) where the boolean is true if the witness was removed or false if the witness was not present, or Hash if the witness was removed and it was the last witness remaining in this tree.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.