penumbra_sdk_tct::internal
Module height

Available on crate feature internal only.

Type level machinery used to statically determine the height of a tree, to ensure the correct domain separator is used when hashing at any height.

Because the height of a tree is inferred by the type system, this means that bugs where the wrong height is used to compute a subtree’s hashing domain separator are greatly reduced.

This module contains type-level code for computing the height of structures and translating an unary representation good for type-level constraints (Succ and Zero) into constant u64s suitable for value-level computation.

Structs

Succ

Height N + 1.

Zero

Height zero.

Traits

Height

Trait identifying the statically-known height of a given tree element.

IsHeight

The constant u8 associated with each unary height.