penumbra_sdk_tct::internal::interface
Trait Complete

pub trait Complete: Height + GetHash {
    type Focus: Focus<Complete = Self>;
}

Available on crate feature internal only.

Marker trait for a type which is the frozen completion of some Focused insertion point.

It is enforced by the type system that Complete and Focus are dual to one another.

Required Associated Types

type Focus: Focus<Complete = Self>

The corresponding Focus of this Complete (i.e. the type which will become this type when it is finalize_owned).