penumbra_sdk_tct::internal::interface
Trait GetPosition

pub trait GetPosition {
    // Required method
    fn position(&self) -> Option<u64>;
}

Available on crate feature internal only.

Get the position of the next insertion into the tree.

Required Methods

fn position(&self) -> Option<u64>

The position of the next insertion into the tree.

Returns None if the tree is full.