penumbra_sdk_tct::internal
Module three

Available on crate feature internal only.

A wrapper around Vec for vectors whose length is at most 3 elements.

This is used in the implementation of frontier::Nodes to store the lefthand siblings of the frontier’s rightmost child, which must number at most 3 (because nodes must have at most 4 children total).

Structs

Three

A vector capable of storing at most 3 elements.

Enums

Elems

All the possible cases of the elements in a Three, by reference.

ElemsMut

All the possible cases of the elements in a Three, by mutable reference.

IntoElems

All the possible cases of the elements in a Three, by value.