penumbra_sdk_tct
Module storage

Incremental serialization and non-incremental deserialization for the Tree.

Re-exports

pub use in_memory::InMemory;

Modules

in_memory

An in-memory storage backend, useful for testing.

Structs

DeleteRange

An update to the underlying storage that constitutes deleting a range of hashes and commitments.

LoadCommitments

Builder for loading commitments to create a Tree.

LoadHashes

Builder for loading hashes to create a Tree.

StoreCommitment

An update to the underlying storage that constitutes storing a single commitment.

StoreHash

An update to the underlying storage that constitutes storing a single hash.

Updates

A collection of updates to the underlying storage.

Enums

StoredPosition

A stored position for the tree: either the position of the tree, or a marker indicating that it is full, and therefore does not have a position.

Update

A single update to the underlying storage, as a data type.

Traits

AsyncRead

An async storage backend capable of reading stored Hashes and [Commitment]s as well as storing the current Position.

AsyncWrite

An async storage backend capable of writing Hashes and [Commitment]s, and garbage-collecting those which have been forgotten.

Read

A synchronous storage backend capable of reading stored Hashes and [Commitment]s as well as storing the current Position.

Write

A synchronous storage backend capable of writing Hashes and [Commitment]s, and garbage-collecting those which have been forgotten.