penumbra_sdk_tct::validate
Function all_proofs

pub fn all_proofs(tree: &Tree) -> Result<(), InvalidWitnesses>

Verify that every witnessed commitment can be used to generate a proof of inclusion which is valid with respect to the current root.

This is an expensive operation that requires traversing the entire tree structure and doing a lot of hashing.

If this ever returns Err, it indicates either a bug in this crate, or a tree that was deserialized from an untrustworthy source.