penumbra_sdk_tct::internal::height
Trait Height

pub trait Height {
    type Height: Path;
}

Available on crate feature internal only.

Trait identifying the statically-known height of a given tree element.

This is used to differentiate the hashes at each level of the tree.

Required Associated Types

type Height: Path

The height of this type above the leaves of the tree.