penumbra_sdk_tct::internal::interface
Trait Focus

pub trait Focus: Height<Height = <Self::Complete as Height>::Height> + GetHash {
    type Complete: Complete<Focus = Self>;

    // Required method
    fn finalize_owned(self) -> Insert<Self::Complete>;
}

Available on crate feature internal only.

A type which can be the focus of an Frontier tree: it can be finalized to make a Complete tree.

Required Associated Types

type Complete: Complete<Focus = Self>

The Complete of this Frontier.

Required Methods

fn finalize_owned(self) -> Insert<Self::Complete>

Transition from an Frontier to being Complete.