penumbra_sdk_tct::internal::path
Trait Path

pub trait Path: IsHeight + Sized {
    type Path;

    // Required method
    fn root(path: &Self::Path, index: u64, leaf: Hash) -> Hash;
}

Available on crate feature internal only.

Identifies the unique type representing an authentication path for the given height.

Required Associated Types

type Path

The authentication path for this height.

Required Methods

fn root(path: &Self::Path, index: u64, leaf: Hash) -> Hash

Calculate the root hash for a path leading to a leaf with the given index and hash.

Object Safety

This trait is not object safe.